+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Removed the deprecated `Get-VCFNsxEdgeCluster` alias from `Get-VCFEdgeCluster`.
 - Removed the deprecated `New-VCFNsxEdgeCluster` alias from `New-VCFEdgeCluster`.
 - Refactors `Invoke-VCFCommand` to run commands on SDDC Manager without the need for SSH across PowerShell editions and operating systems.
+- Added `Export-VCFManagementDomainJsonSpec` to export the JSON spec required for bringup from the Planning & Preparation Excel workbook.
 
 ## v2.4.1